Add remainder of sprint-12 files
